+++ CHANGELOG.md
@@ -7,6 +7,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ## [Unreleased] -->
 
+## [v0.3.0] - 2026-03-26
+
+### Added
+
+* Add parsing of OSC dynamic color sequences from the terminal. ([980077903880](https://github.com/helix-editor/termina/commit/980077903880))
+
+### Fixed
+
+* Catch OSC sequences in the `Event::is_escape` predicate. ([3c1e399b8dd1](https://github.com/helix-editor/termina/commit/3c1e399b8dd1))
+
+## [v0.2.0] - 2026-03-14
+
+### Added
+
+* Implement legacy console API for Windows ([#16](https://github.com/helix-editor/termina/pull/16))
+* Add OSC sequences for changing and resetting terminal colors ([#17](https://github.com/helix-editor/termina/pull/17))
+
+### Fixed
+
+* Fix reading VT data which includes non-ASCII characters from Windows ([#15](https://github.com/helix-editor/termina/pull/15))
+
 ## [v0.1.1] - 2025-09-12
 
 ### Added